Snapshot commit of (broken) dependency and POS-tagging code.

+++ java/src/edu/ohsu/cslu/dep/NivreParserFeatureExtractor.java
@@ -92,7 +92,7 @@ public class NivreParserFeatureExtractor extends FeatureExtractor<NivreParserCon
     }
 
     @Override
-    public int featureCount() {
+    public long featureCount() {
         return featureVectorLength;
     }
